+++ Changelog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 2018-03-23
 ### Changed
 - Updated Documentation.md and README.md
-- Updated indicators.py Busybox version regex search
+- Updated trommel.py, indicators.py, and indicator_config.py to print and write instruction set architecture based on the busybox binary
 
 ## 2018-03-22
 ### Added